Initial integration testing for IREvaluator
This gets a setup where we can point it at a bundle (just one for now)
then evaluate a query with some input document. There are a handful
of negative tests passing right now, they cover the little bit of
logic when initializing an evaluator and starting to process a query).

There are also some failing tests for the "success" cases where we
need to actually perform an evaluation. There are a couple of test
bundles with plans and a TODO list of statements we need to complete
to get them working.

To help with some of these tests plans I threw in a helper script
that can generate the `plan.json` for us from some test bundle.

Signed-off-by: Patrick East <p_east@apple.com>