fix evaluation control-flow to correctly handle scopes (#7)
Previously, the evaluator never popped scopes off the stack and never
continued evaluating statements in parent scopes. This change corrects
this behavior by detecting when we fall off the end of the last block in
a scope, and popping its parent scopes until one is found with
additional statements to evaluate.

Signed-off-by: Oren Shomron <shomron@apple.com>