ast: implement Decodable protocol for RegoValue (#10)
Finally cracked the code on how to make AST.RegoValue conform to Decodable! 🧬
This is part of a a soft-rollout for this functionality - we have some top-level initializers that are still using the JSONSerialization-based decoding logic. We should benchmark the two approaches and see which one is preferable.

Also in a followup - there was some custom decoding logic that can be simplified, e.g. in the conformance tests.

This change also breaks apart AST.RegoValue functionality into several files for clarity.

Signed-off-by: Oren Shomron shomron@apple.com
Signed-off-by: Oren Shomron <shomron@apple.com>