eval: overhaul core evaluation logic 🔧 (#27)
Refactor core evaluation logic to use a dedicated Swift function (evalBlock) for each block, whether it is part of the top-level Plan, BlockStmt, NotStmt or WithStmt.

Previously, block evaluation all happened in-line in evalFrame, and we had to manage our own stack and block/statement pointers manually. This led to several complications, including how to correctly handle BreakStmt and breaking out of several layers of blocks. Additionally, we conflated scopes (and sharing of locals or not) with the context in which each block executed, which led to more issues. Finally, propagation of undefined back up the call stack was confusing.

Now whenever we determine a block is undefined, we can simply return .undefined and handle its propagation at the caller. Additionally, we finally have correct break behavior, which is managed by returning a breakCounter to each caller which they then inspect and respond to.

Compliance tests passing are up from ✅ 771->991 📈

Signed-off-by: Oren Shomron <shomron@apple.com>