An implementation of sprintf builtin (#28)
Probably close enough, but we likely can toil on this for a loong time
to achieve perfect emulation of the Go formatter and upstream OPA.

There is a huge number of compliance tests generated and we're only
passing about half of them. We'll need to keep iterating on this
but it probably isn't worth going too deep right now.

Signed-off-by: Patrick East <p_east@apple.com>