Fixup tracing with IR.AnyStatement (#49)
This required plumbing through some of the "caller" fields as the
enum'd `AnyStatement` rather than as `any Statement`.

It also required a couple of helpers for the tracing code to derive
info about the statement from the enum. We could simplify this if we
get rid of the unknown enum value.. TBD if we want to keep that or
not longer term.

Anyway, we've got per statement tracing back in and showing stuff.

Signed-off-by: Patrick East <p_east@apple.com>