Let input be undefined when it is not specified
I'm not 100% sure this is right, but it seems to match up with jarl
and wasm. We're going to not check for undefined arguments on call
statements unless they are for a builtin. It appears that plans are
expecting to be able to evaluate blocks in called functions.

The main issue being that `input` and `data` are always provided, but
sometimes `input` is undefined. In that case it means all call stmts
start to fail for the plan (not desired behavior).

Signed-off-by: Patrick East <p_east@apple.com>