Support strict builtin errors (#61)
The general pattern being that builtins throw errors when something
is wrong (invalid args, some undefined behavior, failures, etc) and
it is up to the caller to decide out whether or not to propagate the
error.

This introduces a new `BuiltinErrors.halt` flavor of error which, when
thrown, will _always_ propagate and stop the evaluation. These are
to be used for operations that signify something seriously wrong and
the policy evaluation should not continue. This concept is the same in
OPA's topdown evaluator and builtin definitions. It seems only a few
builtins use it, but it does get used in some places. So far it does
not appear that we have any implemented which need it.

Signed-off-by: Patrick East <p_east@apple.com>