Update Engine interface to use a "prepared query" style API
We had a `prepare` call before on the `Engine`, but to better align
with the OPA Go SDK semantics (and set ourselves up for, hopefully,
a smoother transition to a future world where we have to parse+plan
from Rego source) we're switching to a pattern of moving `evaluate`
to only be on a `PreparedQuery`, which is generated by calling the
`prepareForEval` interface of the `Engine`.

We should be in a pretty safe position to refactor how we evaluate,
patch/update bundles, etc with this setup.

Signed-off-by: Patrick East <p_east@apple.com>