ensure query can be evaluated during preparation
Continuing my trend of shoving more validation into
`OPA.Engine.prepareForEvaluation`, this method now checks that the
provided query can be evaluated. This should help catch
misconfigurations earlier in application bootstrapping, without
necessarily having to evaluate something first.

I think it's a bit of a deviation from how query preparation works with
the Go implementation. And I could see this check becoming much more
difficult or impossible to implement as query capabilities expand, so I
won't be upset if this is rejected.

There are already some tests checking for this behavior, but they were
expecting it to happen on `evaluate`. Now the error pops up during
`prepareForEvaluation`, but the tests pass either way.

Signed-off-by: Chris Rice <chrisrice@apple.com>