docs: README wording fix (#25)
'Rego declarative policy' is generally not the phrasing we use
elsewhere, I have updated this to be just 'Rego policies'.

The link to Rego and IR should be more than enough to fill in blanks in
readers understandings.

NOTE: We will also need to get the repo description field updated.

Signed-off-by: Charlie Egan <charlieegan3@users.noreply.github.com>
Co-authored-by: Charlie Egan <charlieegan3@users.noreply.github.com>