docs: README improvements
Based on my own noob experience trying to run the code from
the examples provided here last night.

- Expand on prerequisite steps without assuming prior knowledge
- Make example complete - i.e. can copy-paste and run
- Build failure due to missing `platforms` definition
- Add community support footer to better engage community

Note that the code in the example now prepares in a separate
function, as I thought this better demonstrated the reusability
of the returned prepared query. I have no idea whether this is
idiomatic or not though, so happy to change if not.

Signed-off-by: Anders Eknert <anders@eknert.com>