CI: add basic workflow (#27)
This adds a very basic CI workflow, running tests and compliance-tests.
Compliance tests currently don't all pass, so that's non-fatal.

The workflow will create a summary of the differences of passed/failed
compliance tests between main and the PR branch.

Notable side-quests: 
* ci: make test-compliance non-failure
* ComplianceSuite/Makefile: don't require rq and gsed

That'll make it run without a setup step on the Github runners.

* ci: use xcode-select method for swift version

   I have tried various other methods (docker image, setup-swift action),
   but each had its own problems. We can revisit this when setup-swift
   supports 6.2, or when this project can run its compliance tests on Linux.
   (The latter is an issue tracked elsewhere.)

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>